requirements.txt
flask==2.3.2
pymssql==2.2.7
pandas==2.0.3
openpyxl==3.1.2
scikit-learn==1.3.0
python-dotenv==1.0.0